backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0

# Google Cloud
google-cloud-bigquery==3.25.0
google-cloud-firestore==2.18.0
google-cloud-aiplatform==1.68.0

# Neo4j
neo4j==5.24.0

# Async
httpx==0.27.0
aiofiles==24.1.0

# CORS
python-multipart==0.0.9
